[mod] #102 home data layer 로직

+++ app/src/main/java/sopt/motivoo/data/model/request/home/RequestMissionTodayDto.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RequestMissionTodayDto(
     @SerialName("mission_id")
-    val missionId: Long,
+    val missionId: Int,
 )
